+++ Scripts/Expansion/EJ/Rising Tide/MannedCannon.cs
@@ -85,7 +85,7 @@ namespace Server.Items
                         if (shoot)
                         {
                             DoShootEffects();
-                            TimeSpan delay = TimeSpan.FromSeconds((double)t.Range / 10.0);
+                            TimeSpan delay = TimeSpan.FromSeconds(t.Range / 10.0);
 
                             Timer.DelayCall(delay, new TimerStateCallback(OnShipHit), new object[] { (BaseGalleon)t.Entity, t.Location, AmmoType });
                         }
@@ -99,7 +99,7 @@ namespace Server.Items
                         if (shoot)
                         {
                             DoShootEffects();
-                            TimeSpan delay = TimeSpan.FromSeconds((double)t.Range / 10.0);
+                            TimeSpan delay = TimeSpan.FromSeconds(t.Range / 10.0);
 
                             Timer.DelayCall(delay, new TimerStateCallback(OnMobileHit), new object[] { m, t.Location, AmmoType });
                         }
@@ -145,7 +145,7 @@ namespace Server.Items
                 if (LateralOffset > 1 && currentRange % LateralOffset == 0)
                     lateralOffset++;
 
-                TimeSpan delay = TimeSpan.FromSeconds((double)currentRange / 10.0);
+                TimeSpan delay = TimeSpan.FromSeconds(currentRange / 10.0);
 
                 switch (AmmoType)
                 {
@@ -397,7 +397,7 @@ namespace Server.Items
 
             if (ammoInfo != null)
             {
-                int damage = (int)(Utility.RandomMinMax(ammoInfo.MinDamage, ammoInfo.MaxDamage));
+                int damage = Utility.RandomMinMax(ammoInfo.MinDamage, ammoInfo.MaxDamage);
 
                 if (Operator != null)
                 {
@@ -434,7 +434,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)1);
+            writer.Write(1);
 
             writer.Write(Operator);
             writer.Write(CanFireUnmanned);
@@ -470,7 +470,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0);
+            writer.Write(0);
         }
 
         public override void Deserialize(GenericReader reader)
@@ -495,7 +495,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0);
+            writer.Write(0);
         }
 
         public override void Deserialize(GenericReader reader)
@@ -522,7 +522,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0);
+            writer.Write(0);
         }
 
         public override void Deserialize(GenericReader reader)
