+++ Scripts/Expansion/EJ/StoreBought/BulkOrderBookCover.cs
@@ -223,7 +223,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)1);
+            writer.Write(1);
 
             writer.Write(_UsesRemaining);
             writer.Write((int)_CoverType);
@@ -272,7 +272,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0);
+            writer.Write(0);
         }
 
         public override void Deserialize(GenericReader reader)
