+++ Scripts/Expansion/EJ/TreasuresOfKhaldun/Items/PotionOfGloriousFortune.cs
@@ -40,7 +40,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
@@ -74,7 +74,7 @@ namespace Server.Items
         {
             if (UnderEffects(m))
             {
-                return (int)(Bonus / 100);
+                return Bonus / 100;
             }
 
             return 1;
