+++ Scripts/Expansion/EJ/XxUnsortedxX/HouseCraftables.cs
@@ -271,7 +271,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0); // version
+            writer.Write(0); // version
             writer.Write((int)_Type);
         }
 
@@ -352,7 +352,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
@@ -387,7 +387,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0); // version
+            writer.Write(0); // version
             writer.Write((int)ItemType);
         }
 
@@ -426,7 +426,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0); // version
+            writer.Write(0); // version
             writer.Write((int)ItemType);
         }
 
@@ -491,7 +491,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0);
+            writer.Write(0);
             writer.Write((int)this.Type);
             writer.Write((int)_Resource);
         }
@@ -803,7 +803,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)1);
+            writer.Write(1);
             writer.Write((int)this.Type);
             writer.Write((int)_Resource);
         }
@@ -1046,7 +1046,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)1);
+            writer.Write(1);
             writer.Write((int)this.Type);
             writer.Write((int)_Resource);
         }
