+++ Scripts/Expansion/HS/Consumables/GorgonLens.cs
@@ -180,7 +180,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)2); // version
+            writer.Write(2); // version
             writer.Write((int)m_LenseType);
         }
