+++ Scripts/Expansion/HS/DespiseRevamped/DespiseBosses.cs
@@ -44,7 +44,7 @@ namespace Server.Engines.Despise
         {
             if (mob is PlayerMobile)
             {
-                int chance = ArtifactChance + (int)Math.Min(10, ((PlayerMobile)mob).Luck / 180);
+                int chance = ArtifactChance + Math.Min(10, ((PlayerMobile)mob).Luck / 180);
 
                 if (chance >= Utility.Random(100))
                 {
@@ -76,7 +76,7 @@ namespace Server.Engines.Despise
             base.AlterMeleeDamageTo(to, ref damage);
 
             if (m_Wisp != null && !m_Wisp.Deleted && m_Wisp.Alive)
-                damage += (int)((double)damage * WispScalar);
+                damage += (int)(damage * WispScalar);
         }
 
         public override void AlterMeleeDamageFrom(Mobile from, ref int damage)
@@ -84,7 +84,7 @@ namespace Server.Engines.Despise
             base.AlterMeleeDamageFrom(from, ref damage);
 
             if (m_Wisp != null && !m_Wisp.Deleted && m_Wisp.Alive)
-                damage -= (int)((double)damage * WispScalar);
+                damage -= (int)(damage * WispScalar);
         }
 
         public override void AlterSpellDamageTo(Mobile to, ref int damage)
@@ -92,7 +92,7 @@ namespace Server.Engines.Despise
             base.AlterSpellDamageTo(to, ref damage);
 
             if (m_Wisp != null && !m_Wisp.Deleted && m_Wisp.Alive)
-                damage += (int)((double)damage * WispScalar);
+                damage += (int)(damage * WispScalar);
         }
 
         public override void AlterSpellDamageFrom(Mobile from, ref int damage)
@@ -100,7 +100,7 @@ namespace Server.Engines.Despise
             base.AlterSpellDamageFrom(from, ref damage);
 
             if (m_Wisp != null && !m_Wisp.Deleted && m_Wisp.Alive)
-                damage -= (int)((double)damage * WispScalar);
+                damage -= (int)(damage * WispScalar);
         }
 
         public override void OnThink()
@@ -149,7 +149,7 @@ namespace Server.Engines.Despise
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0);
+            writer.Write(0);
             writer.Write(m_Wisp);
         }
 
@@ -236,7 +236,7 @@ namespace Server.Engines.Despise
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0);
+            writer.Write(0);
         }
 
         public override void Deserialize(GenericReader reader)
@@ -320,7 +320,7 @@ namespace Server.Engines.Despise
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0);
+            writer.Write(0);
         }
 
         public override void Deserialize(GenericReader reader)
@@ -399,7 +399,7 @@ namespace Server.Engines.Despise
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0);
+            writer.Write(0);
         }
 
         public override void Deserialize(GenericReader reader)
@@ -478,7 +478,7 @@ namespace Server.Engines.Despise
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0);
+            writer.Write(0);
         }
 
         public override void Deserialize(GenericReader reader)
