+++ Scripts/Expansion/HS/ExploringTheDeep/Items/WinchAssembly.cs
@@ -155,7 +155,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0); // version
+            writer.Write(0); // version
 
             writer.Write(m_flywheel);
             writer.Write(m_wirespool);
@@ -229,7 +229,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0); // version
+            writer.Write(0); // version
 
             writer.Write(m_WinchAssembly);
             writer.Write(m_hatch);
